test(task-graph): extract the streaming task-output repository contract
The Sqlite, Postgres and Supabase streaming repository tests were three
~144-line files differing by 6 and 9 lines respectively: the cache-ref URI
scheme, a foreign scheme to prove foreign refs are declined, and how a
sibling instance over the same backing store is constructed. Everything
else — the ten behavioural assertions — was duplicated three times, so a
fix to any of them had to be applied three times or silently wasn't.

Those three axes are now options on a contract suite shipped from
`@workglow/task-graph/test`, and each backend is a ~40-line shim keeping
only its own setup and teardown. Assertion names and count are unchanged
(30 tests, 10 per backend).

The contract reaches `getStreamPortCodec` / `makeCacheRef` /
`streamRefViaBacking` by package specifier rather than relatively: `./test`
is its own `--packages=external` bundle, and a relative import would inline
a second copy of the codec registry, so a ref encoded by the code under
test would be read back through a different registry instance.

`StreamingIndexedDbTaskOutputRepository` keeps its own suite — it is a
superset (materializing reader, path-traversal rejection, replay parity),
not another instance of this contract.
