+++ packages/test/src/test/task-graph-output-cache/IndexedDbTaskOutputRepository.integration.test.ts
@@ -4,17 +4,17 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
+import { runTaskOutputRepositoryContract } from "@workglow/task-graph/test";
 import { setLogger, uuid4 } from "@workglow/util";
 import { getTestingLogger } from "@workglow/util/test";
 import "fake-indexeddb/auto";
 import { describe } from "vitest";
 import { IndexedDbTaskOutputRepository } from "../../binding/IndexedDbTaskOutputRepository";
-import { runGenericTaskOutputRepositoryTests } from "./genericTaskOutputRepositoryTests";
 
 describe("IndexedDbTaskOutputRepository", () => {
   let logger = getTestingLogger();
   setLogger(logger);
-  runGenericTaskOutputRepositoryTests(async () => {
+  runTaskOutputRepositoryContract(async () => {
     const id = uuid4().replace(/-/g, "_");
     const dbName = `idx_test_${id}`;
     return new IndexedDbTaskOutputRepository(dbName);
