+++ packages/test/src/test/task-graph-output-cache/SqliteTaskOutputRepository.integration.test.ts
@@ -5,17 +5,17 @@
  */
 
 import { Sqlite } from "@workglow/sqlite/storage";
+import { runTaskOutputRepositoryContract } from "@workglow/task-graph/test";
 import { setLogger, uuid4 } from "@workglow/util";
 import { getTestingLogger } from "@workglow/util/test";
 import { describe } from "vitest";
 import { SqliteTaskOutputRepository } from "../../binding/SqliteTaskOutputRepository";
-import { runGenericTaskOutputRepositoryTests } from "./genericTaskOutputRepositoryTests";
 
 describe("SqliteTaskOutputRepository", async () => {
   await Sqlite.init();
   let logger = getTestingLogger();
   setLogger(logger);
-  runGenericTaskOutputRepositoryTests(async () => {
+  runTaskOutputRepositoryContract(async () => {
     const id = uuid4().replace(/-/g, "_");
     const dbName = `task_output_test_${id}`;
     return new SqliteTaskOutputRepository(":memory:", dbName);
